Update client to new SDK structure.
