+++ .gitignore
@@ -1,9 +1,5 @@
 # Node.js folders
-*coverage/
-*node_modules/
-
-# Build artefacts
-*build/
+node_modules/
 
 # OS generated files and folders
 .DS_Store
