--- client/index.css
@@ -1,78 +1,0 @@
-body, input, button {
-  font-family: Arial, sans-serif;
-  font-size: 18px;
-}
-
-.screen {
-  position: absolute;
-  top: 0px;
-  right: 0px;
-  bottom: 0px;
-  left: 0px;
-  display: flex;
-  flex-direction: column;
-}
-
-.screen > * {
-  padding: 20px;
-}
-
-.messages {
-  flex: 1 1 100%;
-  margin: 0px;
-  overflow: scroll;
-}
-
-.messages .message {
-  list-style-type: none;
-  border-bottom: 1px solid #e8e8e8;
-  line-height: 1.8;
-  padding: 12px 0 20px 0;
-}
-
-.messages .message:first-child {
-  padding-top: 0;
-}
-
-.messages .message:last-child {
-  padding-bottom: 0;
-  border-bottom: none;
-}
-
-.messages .message .likes {
-  display: inline-block;
-  padding: 0px 6px;
-  font-size: 14px;
-  border: 1px solid #cacaca;
-  border-radius: 2px;
-}
-
-.messages .message .likes > * {
-  pointer-events: none;
-}
-
-.messages .message .likes:hover {
-  background: #eee;
-  cursor: pointer;
-}
-
-.send-message-form {
-  display: flex;
-  flex: 0 0 auto;
-  background: #404040;
-}
-
-.send-message-form input {
-  flex: 1 1 100%;
-  margin-right: 20px;
-  padding: 6px 8px;
-}
-
-.send-message-form button {
-  flex: 0 0 auto;
-  border: none;
-  color: #fff;
-  padding: 8px 10px;
-  border-radius: 2px;
-  background: #29abe2;
-}
