--- client/index.html
@@ -1,26 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8">
-
-    <title>Chat</title>
-
-    <link rel="stylesheet" href="/index.css" />
-  </head>
-
-  <body>
-    <div class="screen">
-      <ul class="messages"></ul>
-
-      <form class="send-message-form">
-        <input class="new-message" type="text" placeholder="Enter a message" />
-        <button type="submit">Send message</button>
-      </form>
-    </div>
-
-    <script type="text/javascript" src="/vendor/wolkenkit-client.browser.min.js"></script>
-
-    <script type="text/javascript" src="/view.js"></script>
-    <script type="text/javascript" src="/index.js"></script>
-  </body>
-</html>
