+++ client/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "wolkenkit-template-chat-client",
+  "version": "1.0.0",
+  "scripts": {
+    "build": "webpack --mode production",
+    "serve": "webpack-serve --open"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "html-webpack-plugin": "3.2.0",
+    "webpack": "4.12.0",
+    "webpack-cli": "3.0.7",
+    "webpack-serve": "1.0.4",
+    "wolkenkit-client": "1.2.0"
+  }
+}
