+++ client/src/index.html
@@ -1,0 +1,91 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8">
+
+    <title>Chat</title>
+
+    <style>
+      body, input, button {
+        font-family: Arial, sans-serif;
+        font-size: 18px;
+      }
+      .screen {
+        position: absolute;
+        top: 0px;
+        right: 0px;
+        bottom: 0px;
+        left: 0px;
+        display: flex;
+        flex-direction: column;
+      }
+      .screen > * {
+        padding: 20px;
+      }
+      .messages {
+        flex: 1 1 100%;
+        margin: 0px;
+        overflow: scroll;
+      }
+      .messages .message {
+        list-style-type: none;
+        border-bottom: 1px solid #e8e8e8;
+        line-height: 1.8;
+        padding: 12px 0 20px 0;
+      }
+      .messages .message:first-child {
+        padding-top: 0;
+      }
+      .messages .message:last-child {
+        padding-bottom: 0;
+        border-bottom: none;
+      }
+      .messages .message .likes {
+        display: inline-block;
+        padding: 0px 6px;
+        font-size: 14px;
+        border: 1px solid #cacaca;
+        border-radius: 2px;
+        user-select: none;
+      }
+      .messages .message .likes > * {
+        pointer-events: none;
+      }
+      .messages .message .likes:hover {
+        background: #eee;
+        cursor: pointer;
+      }
+      .send-message-form {
+        display: flex;
+        flex: 0 0 auto;
+        background: #404040;
+      }
+      .send-message-form input {
+        flex: 1 1 100%;
+        margin-right: 20px;
+        padding: 6px 8px;
+      }
+      .send-message-form button {
+        flex: 0 0 auto;
+        border: none;
+        color: #fff;
+        padding: 8px 10px;
+        border-radius: 2px;
+        background: #29abe2;
+      }
+    </style>
+  </head>
+
+  <body>
+    <div class="screen">
+      <ul class="messages"></ul>
+
+      <form class="send-message-form">
+        <input class="new-message" type="text" placeholder="Enter a message" />
+        <button type="submit">Send message</button>
+      </form>
+    </div>
+
+    <!-- All JavaScript will be injected by webpack here -->
+  </body>
+</html>
